% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{gmean}
\alias{gmean}
\title{geometric mean function}
\usage{
gmean(data, ...)
}
\arguments{
\item{data}{vector with the data}

\item{...}{further arguments passed to or from other methods}
}
\value{
geometric mean value
}
\description{
Function to compute the geometric mean
}
\examples{
gmean(1:10)

}
