% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{number_of_na}
\alias{number_of_na}
\title{number of Na's}
\usage{
number_of_na(data, ...)
}
\arguments{
\item{data}{vector with the data}

\item{...}{further arguments passed to or from other methods}
}
\value{
number of Na's
}
\description{
Function to compute the number of Na's
}
\examples{
number_of_na(1:10)

}
