% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{performance_table}
\alias{performance_table}
\title{Table function}
\usage{
performance_table(data, tolatex = FALSE)
}
\arguments{
\item{data}{data frame with the data}

\item{tolatex}{true to print the latex code of the table}
}
\value{
the data frame with the metrics
}
\description{
Function to generate a table with different metrics
}
\examples{
example_data = generate_random_example(seed = 1234)
table(example_data)

}
