% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{q1}
\alias{q1}
\title{1st quartile function}
\usage{
q1(data, ...)
}
\arguments{
\item{data}{vector with the data}

\item{...}{further arguments passed to or from other methods}
}
\value{
1st quartile value
}
\description{
Function to compute the first quartile
}
\examples{
q1(1:10)

}
