% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{q3}
\alias{q3}
\title{3st quartile function}
\usage{
q3(data, ...)
}
\arguments{
\item{data}{vector with the data}

\item{...}{further arguments passed to or from other methods}
}
\value{
3st quartile value
}
\description{
Function to compute the third quartile
}
\examples{
q3(1:10)

}
